Move ZooKeeperRetries.h to Common
